requirements.txt
Flask
mail-parser
ipwhois
extract_msg
whois
pythonwhois-alt
tldextract
homoglyphs
pandas
nltk
sklearn
plotly

matplotlib
seaborn